Match format of strcase
One of the common, default functions we provide is annotated with `+mint:transform:date_in_utc`. The function we use to turn that annotation into a library call tries to call the function `mint.DateInUtc` because of how it turns snakecase into camelcase: that code has no concept of initialisms and so has no idea that UTC should be in caps.

Rather than write some kind of messy lookup table (which we may end up doing anyway one day), it seems more sensible to rename this function. Certainly since nobody uses this, because it is still in development