Ensure collections of complex types marshall with no initialiser
The initialiser we generate is largely only useful for nil values;
specifically when we're building unmarshallers.

It just so happens that it also works for marshalling builtin/ native types.

For complex types, the method of passing a value to `new()` wont work; that
isn't what `new()` does.

Instead, though, because our `mint.MarshallerUnmarshallerValuer` expects a pointer
receiver anyway, we can sub-out intialiser functions for an ampersand.