Set empty collections to nil value
Given a type:

```golang
type Foo struct {
     Bar []string
}
```

When `Foo.Bar` is uninitialised/ `nil`, we Marshall with a Len of
0. When we then Unmarshall we initialise a `[]string` of length 0.

The problem here is that a `nil` slice and a slice of `len()` 0 are two
different things. Thus, when we try:

```golang
in := new(Foo)
in.Marshall(buf)

out := new(Foo)
out.Unmarshall(buf)
```

our types don't match up.

By testing whether `f.Len() == 0` and setting `Foo.Bar` as the nil
value, rather than anything else, we know our collection should be nil.

Of course, this wont work if the collection _should_ be of length 0.